Credits are now updated when you gain them for completing a job
Apparently, it was just the UI wasn't updating, but I didn't realize that so I created a new system for it.  The new system should make it easier for rooms such as the brig and storage containers to increase the payout like they say they do.  Just call the AddPayout function in ShipStats.