Updated the way the game checks for death
Now when you die as a result of an event choice, you go to the death screen as soon as time is unpaused, or as soon as the choice is made if hasSubsequentChoices is set to true.  Before this update, it only checked for death each tick event.  It still checks each tick event just in case.