More progress on adding character Events
New eventsystem algorithms for choosing characterEvent

New pause function in the travel functions

Functions in CharacterEvents to change stats and show the stat change

New variable for outcomes for whether or not this is a correct answer

To do: add a new list for possible character events. possibly to the room itself?