Lots of Minor fixes
Lots of minor fixes around trying to get the UI to show the right number. Had to make a change that made it ugly but needed otherwise the stat change UI didn't popup. The stat change UI needs to change to showcase how much the value changed vs the new total. Might need to change how it's displayed when changing output of a room based on crew.