Added Option to have stat change UI on rooms show opposite -/+ compared to stat bar
Noticed that when add/subtracting crew from rooms it showed subtracting crew on the room when adding because it was subtracting crew from the crew unassigned. Thus, I added an option to show the opposite -/+ compared to the stat bar. So, when adding crew to a room it will show crew being added to the room and subtracted from the crew unassigned in the stat bar.