Removed Change State from Choices and Updated Which Game States stuff was going to due to removing Crew Management
Removed Change Gmae State from Choice Outcomes as it wasn't needed and probably causing bugs with the end of the game as there are other parts of the code that take care of transitioning to the endings.

Fixed other scenes that had their Game State transitions messed up due to Crew Management being removed as an InGameState.

Fixed Payment in Stat bar wrapping by disabling wrapping and enabling overflow.